Crate navigator

This is experimental cross-platform navigation for Signer. Ideally it should replace almost everything and become the only interface

Modules

alerts

List of all alerts

modals

List of all modals

screens

List of all screens

Structs

STATE

Navigation state of the app

Enums

Action

All actions

Error

Functions

do_action

User actions handler.

export_key_info

Export key info with derivations.

export_signatures_bulk

Export signatures bulk.

init_navigation

Should be called in the beginning to recall things stored only by phone

keys_by_seed_name

Get keys by seed name

update_seed_names

Should be called when seed names are modified in native to synchronize data

Type Definitions

Result